Fix recording thread crash on clipped audio samples; bump version to 0.1.17
CalculatePeak called Math.Abs on a 16-bit sample. A fully clipped sample
(-32768) throws OverflowException, which kills NAudio's record thread mid-
recording. The recording was then lost while the UI still showed
"processing", and nothing reached the last-text fallback. Compute the
absolute value on a 32-bit int instead.

+++ src/Blitztext.Platform.Windows/WindowsAudioRecorder.cs
@@ -117,7 +117,7 @@ public sealed class WindowsAudioRecorder : IAudioRecorder, IDisposable
         var peak = 0;
         for (var index = 0; index + 1 < bytesRecorded; index += 2)
         {
-            var sample = Math.Abs(BitConverter.ToInt16(buffer, index));
+            var sample = Math.Abs((int)BitConverter.ToInt16(buffer, index));
             if (sample > peak)
             {
                 peak = sample;
